修改acceptors/list
