+++ src/routes/acceptors/list/index.js
@@ -1,7 +1,9 @@
 import React from 'react';
 import ListAcceptors from './ListAcceptors';
-import { fetchAcceptors, cleanAcceptors } from '../../actions/list-acceptors';
-import fetch from '../../core/fetch';
+import { fetchAcceptors } from '../../../actions/acceptors/list';
+import { reset } from '../../../actions/common';
+
+// import fetch from '../../../core/fetch';
 
 export default {
 
@@ -9,16 +11,16 @@ export default {
 
   async action({ query, context }) { // eslint-disable-line react/prop-types
     const { dispatch } = context.store;
-    const qs = Object.assign({}, {
+    const qs = Object.assign({ text: '' }, {
       pageIndex: 0,
       pageSize: 20,
     }, query);
-    dispatch(cleanAcceptors());
+    dispatch(reset());
     dispatch(fetchAcceptors(query));
 
     const props = {
       fetchAcceptors,
-      cleanAcceptors,
+      cleanAcceptors: reset,
       query: qs,
     };
     return <ListAcceptors {...props} />;
